# Sort mappings on-demand to greatly improve performance