--- a/lib/source-map-consumer.js
+++ b/lib/source-map-consumer.js
@@ -144,8 +144,11 @@ SourceMapConsumer.prototype.eachMapping =
 
     var sourceRoot = this.sourceRoot;
     mappings.map(function (mapping) {
-      var source = mapping.source === null ? null : this._sources.at(mapping.source);
-      source = util.computeSourceURL(sourceRoot, source, this._sourceMapURL);
+      var source = null;
+      if(mapping.source !== null) {
+        source = this._sources.at(mapping.source);
+        source = util.computeSourceURL(sourceRoot, source, this._sourceMapURL);
+      }
       return {
         source: source,
         generatedLine: mapping.generatedLine,