--- a/lib/source-map-consumer.js
+++ b/lib/source-map-consumer.js
@@ -67,7 +67,7 @@ Object.defineProperty(SourceMapConsumer.prototype, '_generatedMappings', {
   enumerable: true,
   get: function () {
     if (!this.__generatedMappings) {
-      this._parseMappings(this._mappings, this.sourceRoot);
+      this._sortGeneratedMappings();
     }
 
     return this.__generatedMappings;
@@ -80,10 +80,36 @@ Object.defineProperty(SourceMapConsumer.prototype, '_originalMappings', {
   enumerable: true,
   get: function () {
     if (!this.__originalMappings) {
-      this._parseMappings(this._mappings, this.sourceRoot);
+      this._sortOriginalMappings();
     }
 
     return this.__originalMappings;
+  }
+});
+
+SourceMapConsumer.prototype.__generatedMappingsUnsorted = null;
+Object.defineProperty(SourceMapConsumer.prototype, '_generatedMappingsUnsorted', {
+  configurable: true,
+  enumerable: true,
+  get: function () {
+    if (!this.__generatedMappingsUnsorted) {
+      this._parseMappings(this._mappings, this.sourceRoot);
+    }
+
+    return this.__generatedMappingsUnsorted;
+  }
+});
+
+SourceMapConsumer.prototype.__originalMappingsUnsorted = null;
+Object.defineProperty(SourceMapConsumer.prototype, '_originalMappingsUnsorted', {
+  configurable: true,
+  enumerable: true,
+  get: function () {
+    if (!this.__originalMappingsUnsorted) {
+      this._parseMappings(this._mappings, this.sourceRoot);
+    }
+
+    return this.__originalMappingsUnsorted;
   }
 });
 
@@ -101,6 +127,20 @@ SourceMapConsumer.prototype._charIsMappingSeparator =
 SourceMapConsumer.prototype._parseMappings =
   function SourceMapConsumer_parseMappings(aStr, aSourceRoot) {
     throw new Error("Subclasses must implement _parseMappings");
+  };
+
+SourceMapConsumer.prototype._sortGeneratedMappings =
+  function SourceMapConsumer_sortGeneratedMappings() {
+    const mappings = this._generatedMappingsUnsorted;
+    quickSort(mappings, util.compareByGeneratedPositionsDeflated);
+    this.__generatedMappings = mappings;
+  };
+
+SourceMapConsumer.prototype._sortOriginalMappings =
+  function SourceMapConsumer_sortOriginalMappings() {
+    const mappings = this._originalMappingsUnsorted;
+    quickSort(mappings, util.compareByOriginalPositions);
+    this.__originalMappings = mappings;
   };
 
 SourceMapConsumer.GENERATED_ORDER = 1;
@@ -564,11 +604,9 @@ BasicSourceMapConsumer.prototype._parseMappings =
       }
     }
 
-    quickSort(generatedMappings, util.compareByGeneratedPositionsDeflated);
-    this.__generatedMappings = generatedMappings;
+    this.__generatedMappingsUnsorted = generatedMappings;
 
-    quickSort(originalMappings, util.compareByOriginalPositions);
-    this.__originalMappings = originalMappings;
+    this.__originalMappingsUnsorted = originalMappings;
   };
 
 /**
@@ -1097,8 +1135,8 @@ IndexedSourceMapConsumer.prototype.generatedPositionFor =
  */
 IndexedSourceMapConsumer.prototype._parseMappings =
   function IndexedSourceMapConsumer_parseMappings(aStr, aSourceRoot) {
-    this.__generatedMappings = [];
-    this.__originalMappings = [];
+    const generatedMappings = this.__generatedMappingsUnsorted = [];
+    const originalMappings = this.__originalMappingsUnsorted = [];
     for (var i = 0; i < this._sections.length; i++) {
       var section = this._sections[i];
       var sectionMappings = section.consumer._generatedMappings;
@@ -1134,15 +1172,12 @@ IndexedSourceMapConsumer.prototype._parseMappings =
           name: name
         };
 
-        this.__generatedMappings.push(adjustedMapping);
+        generatedMappings.push(adjustedMapping);
         if (typeof adjustedMapping.originalLine === 'number') {
-          this.__originalMappings.push(adjustedMapping);
+          originalMappings.push(adjustedMapping);
         }
       }
     }
-
-    quickSort(this.__generatedMappings, util.compareByGeneratedPositionsDeflated);
-    quickSort(this.__originalMappings, util.compareByOriginalPositions);
   };
 
 exports.IndexedSourceMapConsumer = IndexedSourceMapConsumer;