--- a/test/util.js
+++ b/test/util.js
@@ -249,6 +249,13 @@ exports.emptyMap = {
   sources: [],
   mappings: ''
 };
+exports.mapWithSourcelessMapping = {
+  version: 3,
+  file: 'example.js',
+  names: [],
+  sources: ['example.js'],
+  mappings: 'AAgCA,C'
+};
 
 
 function assertMapping(generatedLine, generatedColumn, originalSource,