The current implementation of `SourceMapConsumer.prototype._parseMappings` eagerly sorts both generated and original mappings.  However, in certain situations, this is unnecessary.  For example, when composing 2 sourcemaps, I do something like this:

```
const jsToMinjsMap = new SourceMapConsumer(map1);
const generator = SourceMapGenerator.fromSourceMap(jsToMinjsMap);
generator.applySourceMap(new SourceMapConsumer(map2);
```

For both consumers, only the generated mappings are used, so the original mappings never need to be sorted.  This saves a big chunk of time.

Here are benchmarking results before and after the changes in this PR.  The "parse" benchmark accesses generated but not original mappings, which matches what happens in the example outlined above.

This PR is based off of #299 because benchmarks are broken on master.

#### Benchmark Before

> Parsing source map
> iteration: 4360.786ms
> iteration: 4122.001ms
> iteration: 4293.866ms
> iteration: 4114.573ms
> iteration: 4317.637ms
> iteration: 4068.610ms
> [Stats samples: 6, total: 25297 ms, mean: 4216.166666666667 ms, stddev: 137.68384073666743 ms]
> 
> Serializing source map
> iteration: 2061.736ms
> iteration: 2000.969ms
> iteration: 2161.078ms
> iteration: 2048.509ms
> iteration: 1940.526ms
> iteration: 2452.196ms
> iteration: 2112.322ms
> iteration: 2751.964ms
> iteration: 3024.207ms
> iteration: 3088.927ms
> iteration: 6230.904ms
> [Stats samples: 11, total: 29878 ms, mean: 2716.181818181818 ms, stddev: 1293.2112356455923 ms]
> 

#### Benchmark After

> Parsing source map
> iteration: 2260.933ms
> iteration: 2450.733ms
> iteration: 2124.171ms
> iteration: 2432.659ms
> iteration: 2404.654ms
> iteration: 3115.290ms
> iteration: 3107.850ms
> iteration: 3031.653ms
> iteration: 2939.054ms
> [Stats samples: 9, total: 23843 ms, mean: 2649.222222222222 ms, stddev: 415.21214457190433 ms]
> 
> Serializing source map
> iteration: 1929.926ms
> iteration: 1919.806ms
> iteration: 1899.222ms
> iteration: 1858.227ms
> iteration: 1845.405ms
> iteration: 2060.502ms
> iteration: 2000.878ms
> iteration: 2274.669ms
> iteration: 2494.753ms
> iteration: 2731.931ms
> iteration: 5644.601ms
> [Stats samples: 11, total: 26655 ms, mean: 2423.181818181818 ms, stddev: 1160.6862711344525 ms]